Change up heavy worker to recognize new task in new module
