+++ backend/onyx/background/celery/versioned_apps/heavy.py
@@ -4,15 +4,11 @@ fetch. Port over the code in those files if we add an EE version of this worker.
 
 from celery import Celery
 
+from onyx.utils.variable_functionality import fetch_versioned_implementation
 from onyx.utils.variable_functionality import set_is_ee_based_on_env_variable
 
 set_is_ee_based_on_env_variable()
-
-
-def get_app() -> Celery:
-    from onyx.background.celery.apps.heavy import celery_app
-
-    return celery_app
-
-
-app = get_app()
+app: Celery = fetch_versioned_implementation(
+    "onyx.background.celery.apps.heavy",
+    "celery_app",
+)
