+++ backend/onyx/background/celery/versioned_apps/primary.py
@@ -7,5 +7,6 @@ from onyx.utils.variable_functionality import set_is_ee_based_on_env_variable
 
 set_is_ee_based_on_env_variable()
 app: Celery = fetch_versioned_implementation(
-    "onyx.background.celery.apps.primary", "celery_app"
+    "onyx.background.celery.apps.primary",
+    "celery_app",
 )
